Octahedrons follow f ( x, y), member octahedrons are of type SceneObject*

+++ surface3d.cpp
@@ -97,7 +97,7 @@ void surface3D::drawObject()
 
 }
 
-float surface3D::function(int x, int y)
+float surface3D::function(float x, float y)
 {
     return (sin(x) + cos(y));
 }
